Big changes, fix corrupted and new anchor logic
Huge changes to innerpath, and the general drawing of the smart shape guides. Including a new way to look at the anchor point and better code for drawing it. It will now also remove itself intelligently. A bigger part of the code now uses the expression values as driver, but it does seem I could add one more value to the ffx that will do this padding calculation for me, since i have to re-reference it so many times.