Fixes, cleaning and effectivising
Maybe multiple layers be able to be converted. Streamlined the ffx in the utlities object, which was also reformatted to follow the other js files, and how the objects were written. Also small bug fixes and improved logic. Along with Rectangle Logic being improved to not calculate edges if not necessary thorugh user input in the ffx.